Update logic to share the same Captcha interface
