--- src/FriendlyCaptcha/FriendlyCaptcha.php
@@ -1,139 +1,0 @@
-<?php
-
-/**
- * Class that holds Friendly Captcha check.
- *
- * @package EightshiftForms\FriendlyCaptcha
- */
-
-declare(strict_types=1);
-
-namespace EightshiftForms\FriendlyCaptcha;
-
-use EightshiftForms\Hooks\Variables;
-use EightshiftForms\Labels\LabelsInterface;
-use EightshiftForms\Exception\BadRequestException;
-use EightshiftForms\Helpers\SettingsHelpers;
-use EightshiftForms\Rest\Routes\AbstractBaseRoute;
-use EightshiftForms\Troubleshooting\SettingsFallback;
-
-/**
- * FriendlyCaptcha class.
- */
-class FriendlyCaptcha implements FriendlyCaptchaInterface
-{
-	/**
-	 * Instance variable of LabelsInterface data.
-	 *
-	 * @var LabelsInterface
-	 */
-	protected $labels;
-
-	/**
-	 * Create a new instance that injects classes.
-	 *
-	 * @param LabelsInterface $labels Inject labels methods.
-	 */
-	public function __construct(LabelsInterface $labels)
-	{
-		$this->labels = $labels;
-	}
-
-	/**
-	 * Check Friendly Captcha request.
-	 *
-	 * @param string $response Response token from frontend.
-	 * @param array<string, mixed> $formDetails Form details.
-	 *
-	 * @throws BadRequestException If captcha is not valid.
-	 *
-	 * @return array<mixed>
-	 */
-	public function check(string $response, array $formDetails = []): array
-	{
-		if (!\apply_filters(SettingsFriendlyCaptcha::FILTER_SETTINGS_GLOBAL_IS_VALID_NAME, false)) {
-			return [
-				AbstractBaseRoute::R_MSG => $this->labels->getLabel('friendlyCaptchaSuccess'),
-				AbstractBaseRoute::R_DEBUG => [
-					AbstractBaseRoute::R_DEBUG_KEY => SettingsFallback::SETTINGS_FALLBACK_FLAG_FRIENDLY_CAPTCHA_FEATURE_DISABLED,
-				],
-			];
-		}
-
-		$debug = [
-			'response' => $response,
-			'formDetails' => $formDetails,
-		];
-
-		if (!$response) {
-			// phpcs:disable Eightshift.Security.HelpersEscape.ExceptionNotEscaped
-			throw new BadRequestException(
-				$this->labels->getLabel('friendlyCaptchaBadRequest'),
-				[
-					AbstractBaseRoute::R_DEBUG_KEY => SettingsFallback::SETTINGS_FALLBACK_FLAG_FRIENDLY_CAPTCHA_REQUEST_MISSING_TOKEN,
-					AbstractBaseRoute::R_DEBUG => $debug,
-				]
-			);
-			// phpcs:enable
-		}
-
-		$siteKey = SettingsHelpers::getOptionWithConstant(Variables::getFriendlyCaptchaSiteKey(), SettingsFriendlyCaptcha::SETTINGS_FRIENDLY_CAPTCHA_SITE_KEY);
-		$apiKey = SettingsHelpers::getOptionWithConstant(Variables::getFriendlyCaptchaApiKey(), SettingsFriendlyCaptcha::SETTINGS_FRIENDLY_CAPTCHA_API_KEY);
-
-		$apiResponse = \wp_remote_post(
-			SettingsFriendlyCaptcha::getEndpointUrl(),
-			[
-				'headers' => [
-					'Content-Type' => 'application/json; charset=utf-8',
-					'X-API-Key' => $apiKey,
-				],
-				'data_format' => 'body',
-				'body' => \wp_json_encode([
-					'response' => $response,
-					'sitekey' => $siteKey,
-				]),
-			]
-		);
-
-		// Generic error msg from WP.
-		if (\is_wp_error($apiResponse)) {
-			// phpcs:disable Eightshift.Security.HelpersEscape.ExceptionNotEscaped
-			throw new BadRequestException(
-				$this->labels->getLabel('submitWpError'),
-				[
-					AbstractBaseRoute::R_DEBUG_KEY => SettingsFallback::SETTINGS_FALLBACK_FLAG_FRIENDLY_CAPTCHA_REQUEST_WP_ERROR,
-					AbstractBaseRoute::R_DEBUG => $debug,
-				]
-			);
-			// phpcs:enable
-		}
-
-		// Get body from the response.
-		$responseBody = \json_decode(\wp_remote_retrieve_body($apiResponse), true) ?? [];
-
-		$debug = \array_merge($debug, [
-			'responseBody' => $responseBody,
-		]);
-
-		// Check the success field.
-		if (empty($responseBody['success'])) {
-			// phpcs:disable Eightshift.Security.HelpersEscape.ExceptionNotEscaped
-			throw new BadRequestException(
-				$this->labels->getLabel('friendlyCaptchaError'),
-				[
-					AbstractBaseRoute::R_DEBUG_KEY => SettingsFallback::SETTINGS_FALLBACK_FLAG_FRIENDLY_CAPTCHA_OUTPUT_ERROR,
-					AbstractBaseRoute::R_DEBUG => $debug,
-				]
-			);
-			// phpcs:enable
-		}
-
-		return [
-			AbstractBaseRoute::R_MSG => $this->labels->getLabel('friendlyCaptchaSuccess'),
-			AbstractBaseRoute::R_DEBUG => [
-				AbstractBaseRoute::R_DEBUG_KEY => SettingsFallback::SETTINGS_FALLBACK_FLAG_FRIENDLY_CAPTCHA_SUCCESS,
-				AbstractBaseRoute::R_DEBUG => $debug,
-			],
-		];
-	}
-}
