+++ src/Rest/Routes/General/CaptchaValidateRoute.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
 namespace EightshiftForms\Rest\Routes\General;
 
 use EightshiftForms\Captcha\CaptchaInterface;
+use EightshiftForms\Captcha\SettingsCaptchaProvider;
 use EightshiftForms\Labels\LabelsInterface;
 use EightshiftForms\Helpers\DeveloperHelpers;
 use EightshiftForms\Rest\Routes\AbstractBaseRoute;
@@ -54,6 +55,23 @@ class CaptchaValidateRoute extends AbstractSimpleFormSubmit
 		$this->validator = $validator;
 		$this->labels = $labels;
 		$this->captcha = $captcha;
+	}
+
+	/**
+	 * Register the route only when Google reCAPTCHA is the active provider.
+	 *
+	 * Friendly Captcha solves entirely in the browser and has no pre-submission
+	 * verification endpoint — exposing this route for it would misroute tokens.
+	 *
+	 * @return void
+	 */
+	public function register(): void
+	{
+		if (SettingsCaptchaProvider::getActiveProvider() !== SettingsCaptchaProvider::PROVIDER_GOOGLE) {
+			return;
+		}
+
+		parent::register();
 	}
 
 	/**
